pyproject.toml
[tool.poetry]
name = "itba-tp-final"
version = "0.1.0"
description = ""
authors = ["lucas mattar"]

[tool.poetry.dependencies]
python = "^3.9"
apache-airflow = {extras = ["amazon"], version = "^2.2.4"}
jupyter = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
